refactor: 메뉴 상세 페이지 컴포넌트 분리, 필요한 더미데이터 추가
