+++ src/components/Detail/userDummy.json
@@ -1,0 +1,5 @@
+{
+  "id": 111,
+  "name": "계란이 조아",
+  "profileImageUrl": "https://via.placeholder.com/300x150"
+}
